Fix merge conflicts: Update import and integrate validate_forecast from validate_forecast.py
